Add files via upload
This project is based on the different services provided by Zomato globally. The aim of the project is to 
construct an interactive and user-friendly dashboard in Power bi which would allow users/clients to access 
and acquire important and crucial information about the services provided by Zomato on a global scale.